alphaspread/requirements.txt
bs4
lxml
argparse
requests
googlesearch-python
pydantic
